// src/app/shared/models/report-follow/report-number-win-bid.model.ts
export class ReportNumberWinBid {
    tenderCount: number;
    tenderAmount: number;
    waitForResultCount: number;
    waitForResultAmount: number;
    winningOfBidCount: number;
    winningOfBidAmount: number;
    winningOfBidCompareTenderPer: number;
    winningOfBidCompareTotalBidPer: number;
}